branch boldim deb aytsam ham boladi!!

+++ src/main/java/uz/nonvoyxona/app/model/dto/request/ProductionDTO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class ProductionDTO {
 
     private int branchId;
-    private Long bakerId;
-    private Integer productId;
-    private Integer quantity;
+    private int bakerId;
+    private int productId;
+    private int quantity;
 }
