+++ src/main/java/uz/nonvoyxona/app/repository/DeliveryItemRepo.java
@@ -3,5 +3,5 @@ package uz.nonvoyxona.app.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.nonvoyxona.app.model.DeliveryItem;
 
-public interface DeliveryItemRepo extends JpaRepository<DeliveryItem, Long> {
+public interface DeliveryItemRepo extends JpaRepository<DeliveryItem, Integer> {
 }
