+++ src/main/java/uz/nonvoyxona/app/repository/InStoreItemRepo.java
@@ -3,5 +3,5 @@ package uz.nonvoyxona.app.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.nonvoyxona.app.model.InStoreItem;
 
-public interface InStoreItemRepo extends JpaRepository<InStoreItem, Long> {
+public interface InStoreItemRepo extends JpaRepository<InStoreItem, Integer> {
 }
